+++ src/main/java/com/meoguri/linkocean/domain/bookmark/service/FavoriteServiceImpl.java
@@ -6,7 +6,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.meoguri.linkocean.domain.bookmark.entity.Bookmark;
 import com.meoguri.linkocean.domain.bookmark.persistence.FindBookmarkByIdQuery;
 import com.meoguri.linkocean.domain.profile.entity.Profile;
-import com.meoguri.linkocean.domain.profile.persistence.FindProfileByIdQuery;
+import com.meoguri.linkocean.domain.profile.persistence.command.FindProfileByIdQuery;
 
 import lombok.RequiredArgsConstructor;
