+++ src/main/java/com/meoguri/linkocean/domain/profile/persistence/command/CustomProfileRepository.java
@@ -1,10 +1,10 @@
-package com.meoguri.linkocean.domain.profile.persistence;
+package com.meoguri.linkocean.domain.profile.persistence.command;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 
 import com.meoguri.linkocean.domain.profile.entity.Profile;
-import com.meoguri.linkocean.domain.profile.persistence.dto.ProfileFindCond;
+import com.meoguri.linkocean.domain.profile.persistence.command.dto.ProfileFindCond;
 
 public interface CustomProfileRepository {
