+++ src/main/java/com/meoguri/linkocean/domain/profile/persistence/command/FindProfileByIdQuery.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.profile.persistence;
+package com.meoguri.linkocean.domain.profile.persistence.command;
 
 import static java.lang.String.*;
