+++ src/main/java/com/meoguri/linkocean/domain/profile/persistence/command/dto/ProfileFindCond.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.profile.persistence.dto;
+package com.meoguri.linkocean.domain.profile.persistence.command.dto;
 
 import lombok.Builder;
 import lombok.Getter;
