+++ src/main/java/com/meoguri/linkocean/domain/profile/persistence/query/ProfileQueryRepository.java
@@ -1,0 +1,4 @@
+package com.meoguri.linkocean.domain.profile.persistence.query;
+
+public interface ProfileQueryRepository {
+}
