+++ src/main/java/com/meoguri/linkocean/domain/profile/service/command/ProfileService.java
@@ -3,7 +3,7 @@ package com.meoguri.linkocean.domain.profile.service.command;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 
-import com.meoguri.linkocean.domain.profile.persistence.dto.ProfileFindCond;
+import com.meoguri.linkocean.domain.profile.persistence.command.dto.ProfileFindCond;
 import com.meoguri.linkocean.domain.profile.service.command.dto.GetDetailedProfileResult;
 import com.meoguri.linkocean.domain.profile.service.command.dto.GetProfilesResult;
 import com.meoguri.linkocean.domain.profile.service.command.dto.RegisterProfileCommand;
