+++ src/test/java/com/meoguri/linkocean/domain/profile/persistence/CheckIsFollowQueryTest.java
@@ -13,6 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
 
 import com.meoguri.linkocean.domain.profile.entity.Profile;
+import com.meoguri.linkocean.domain.profile.persistence.command.CheckIsFollowQuery;
 import com.meoguri.linkocean.test.support.persistence.BasePersistenceTest;
 
 @Import(CheckIsFollowQuery.class)
