+++ src/test/java/com/meoguri/linkocean/domain/profile/persistence/FindProfileByIdQueryTest.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
 
 import com.meoguri.linkocean.domain.profile.entity.Profile;
+import com.meoguri.linkocean.domain.profile.persistence.command.FindProfileByIdQuery;
 import com.meoguri.linkocean.test.support.persistence.BasePersistenceTest;
 
 @Import(FindProfileByIdQuery.class)
