+++ src/test/java/com/meoguri/linkocean/domain/profile/persistence/FollowRepositoryTest.java
@@ -12,6 +12,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.meoguri.linkocean.domain.profile.entity.Profile;
+import com.meoguri.linkocean.domain.profile.persistence.command.FollowRepository;
 import com.meoguri.linkocean.test.support.persistence.BasePersistenceTest;
 
 class FollowRepositoryTest extends BasePersistenceTest {
